+++ src/app/task/tasks.core.ts
@@ -339,7 +339,7 @@ export class TasksCore {
             , 'tsk_id_record': task.tsk_id_record || 'general'
             , 'tsk_name': task.tsk_name
             , 'tsk_notes': task.tsk_notes || ''
-            , 'tsk_parent': task.tsk_parent || 0
+            , 'tsk_parent': task.tsk_parent || '0'
             , 'tsk_order': this.nextOrder()
             , 'tsk_date_done': <Date>undefined
             , 'tsk_total_time_spent': 0
@@ -358,7 +358,7 @@ export class TasksCore {
             , 'tsk_template': task.tsk_template || ''
             , 'tsk_template_state': task.tsk_template_state || ''
             , 'tsk_date_due': task.tsk_date_due || <Date>undefined
-            , 'tsk_id_related': task.tsk_id_related || 0
+            , 'tsk_id_related': task.tsk_id_related || '0'
             , 'tsk_url': task.tsk_url || ''
             , 'tsk_ctg_repeats': task.tsk_ctg_repeats || 0
             , 'tsk_id_main': task.tsk_id_main || id
@@ -925,7 +925,7 @@ export class TasksCore {
     compareTask(t: any, s: any){
         //let fields = ['tsk_id_container','tsk_id_record','tsk_name','tsk_notes', 'tsk_parent', 'tsk_order', 'tsk_date_done', 'tsk_total_time_spent', 'tsk_ctg_in_process', 'tsk_qualifiers', 'tsk_tags', 'tsk_estimated_duration', 'tsk_schedule_date_start', 'tsk_schedule_date_end', 'tsk_date_view_until', 'tsk_id_user_added', 'tsk_id_user_asigned', 'tsk_date_add', 'tsk_date_mod', 'tsk_ctg_status'];
         //let fields = ['tsk_id_container','tsk_id_record','tsk_name','tsk_notes','tsk_parent','tsk_order','tsk_date_done','tsk_total_time_spent','tsk_ctg_in_process','tsk_qualifiers','tsk_tags','tsk_estimated_duration','tsk_schedule_date_start','tsk_schedule_date_end','tsk_date_view_until','tsk_id_user_added','tsk_id_user_asigned','tsk_template','tsk_template_state','tsk_date_due','tsk_id_related','tsk_url','tsk_ctg_repeats','tsk_id_main','tsk_ctg_rep_type','tsk_ctg_rep_after_completion','tsk_ctg_rep_end','tsk_rep_date_end','tsk_rep_end_iteration','tsk_rep_iteration','tsk_rep_frequency','tsk_ctg_rep_frequency_rule','tsk_rep_weekdays','tsk_date_add','tsk_date_mod','tsk_ctg_status'];
-        let fields = ['tsk_id_container','tsk_id_record','tsk_name','tsk_notes','tsk_parent','tsk_order','tsk_date_done','tsk_total_time_spent','tsk_ctg_in_process','tsk_tags','tsk_estimated_duration','tsk_schedule_date_start','tsk_schedule_date_end','tsk_date_view_until','tsk_id_user_added','tsk_id_user_asigned','tsk_template','tsk_template_state','tsk_date_due','tsk_id_related','tsk_url','tsk_id_main','tsk_rep_date_end','tsk_rep_weekdays','tsk_date_add','tsk_date_mod','tsk_ctg_status'];
+        let fields = ['tsk_id_container','tsk_id_record','tsk_name','tsk_notes','tsk_parent','tsk_order','tsk_date_done','tsk_total_time_spent','tsk_ctg_in_process','tsk_tags','tsk_estimated_duration','tsk_schedule_date_start','tsk_schedule_date_end','tsk_date_view_until','tsk_id_user_added','tsk_id_user_asigned','tsk_template','tsk_template_state','tsk_date_due','tsk_id_related','tsk_url','tsk_id_main','tsk_rep_date_end','tsk_rep_weekdays','tsk_date_add','tsk_ctg_status'];
         //let fields = ['tsk_date_done'];
         let comparison: Array<any> = [];
         let field: any = {};
@@ -933,10 +933,16 @@ export class TasksCore {
         fields.forEach((f: string) => {
             field = {};
             field.id = t.tsk_id;
+            field.displayName = t.tsk_id + ' - ' + t.tsk_name;
             field.name = f;
             field.client = t[f];
             field.server = s[f];
-            field.isEqual = t[f] == s[f];
+            if (t[f] instanceof Date) {
+                s[f] = new Date(s[f]);
+                field.isEqual = t[f].getTime() == s[f].getTime();
+            } else {
+                field.isEqual = t[f] == s[f];
+            }
             if (field.name === 'tsk_total_time_spent' && !field.isEqual){
                 // let's find out which one is the real one
                 let taskClient: Task = this.data.taskList.find((t: Task) => t.tsk_id === field.id);
@@ -947,7 +953,7 @@ export class TasksCore {
                 // we preserve server date if it defers from browser
                 let taskClient: Task = this.data.taskList.find((t: Task) => t.tsk_id === field.id);
                 taskClient[f] = new Date(s[f]);
-                field.isEqual = taskClient[f] == s[f];
+                field.isEqual = taskClient[f].getTime() == new Date(s[f]).getTime();
             }
             if (!field.isEqual){
                 comparison.push(field);